Convert all id fields in Go backend to int64
- Added ParseableInt64 type with JSON unmarshaling support
- Updated all domain model ID fields from int to int64:
  - Portfolio.Id, Asset.Id, PlannedAllocation.Id, AllocationPlanIdentifier.Id, PortfolioObservationTimestamp.Id
- Updated all repository interface and implementation method signatures
- Updated all domain service method signatures
- Updated all application service method signatures
- Updated API REST models to use ParseableInt64 and int64
- Updated REST controllers to use strconv.ParseInt instead of strconv.Atoi
- Updated REST mapping functions to handle int64 conversions
- Fixed integration test utilities to work with int64 IDs

All Go code compiles successfully and core functionality tests pass.

+++ src/main/go/application/potfolio_allocation_management_service.go
@@ -17,7 +17,7 @@ type PortfolioAllocationManagementAppService struct {
 }
 
 func (service *PortfolioAllocationManagementAppService) MergePortfolioAllocations(
-	portfolioId int,
+	portfolioId int64,
 	observationTimestamp *domain.PortfolioObservationTimestamp,
 	allocations []*domain.PortfolioAllocation,
 ) error {
